updated POST and POST car models, old tests with CarModel, fixed GET tests
